Try to launch inspections
