+++ cli/build.gradle.kts
@@ -3,7 +3,6 @@ version = rootProject.version
 
 dependencies {
     implementation(project(":core"))
-    implementation(project(":plugin"))
 }
 
 tasks {
